+++ ntex-io/CHANGES.md
@@ -2,6 +2,10 @@
 
 ## [0.1.0-b.10] - 2021-12-30
 
+* IoRef::close() method initiates io stream shutdown
+
+* IoRef::force_close() method terminates io stream
+
 * Cleanup Filter trait, removed closed,want_read,want_shutdown methods
 
 * Cleanup internal flags on io error